fix: remover exportação nomeada de página Next.js e corrigir erro de build
