+++ apcpro-web/src/app/dashboard/alunos/page.tsx
@@ -56,37 +56,6 @@ function calcularIdade(dataNascimento?: string): number | undefined {
   return idade >= 0 ? idade : undefined;
 }
 
-// Componente de evolução de massa magra/peso
-type CardEvolucaoProps = {
-  label: string;
-  valor: number;
-  unidade?: string;
-};
-
-export function CardEvolucao({
-  label,
-  valor,
-  unidade = "kg",
-}: CardEvolucaoProps) {
-  const positivo = valor > 0;
-
-  return (
-    <div className="flex items-center gap-2 p-4 bg-white rounded shadow">
-      <span className="font-medium text-gray-700">{label}</span>
-      <span
-        className={`flex items-center gap-1 font-bold ${
-          positivo ? "text-green-600" : "text-red-600"
-        }`}
-      >
-        {positivo && "+"}
-        {valor}
-        {unidade}
-        {positivo && <ArrowUpRight className="w-4 h-4" aria-label="Aumento" />}
-      </span>
-    </div>
-  );
-}
-
 export default function AlunosDashboard() {
   const { profile } = useUserProfile();
   const { proxima, loading } = useProximaAvaliacao(profile?.id ?? ""); // Sempre chamado!
